test(java): split fory json tests
